% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_sqi_mapper}
\alias{data_sqi_mapper}
\title{Quadro de referência para a recodificação dos nomes dos indicadores de supervisão e da chave de filtragem}
\format{
Uma lista das variáveis incluídas no objecto de dados
\describe{
\item{area_responsecode}{area_responsecode}
\item{form_type}{form_type}
\item{domain}{domain}
\item{type}{type}
\item{filter_key}{filter_key}
\item{code}{code}
\item{var_purpose}{var_purpose}
\item{question_description}{question_description}
\item{domain_description}{domain_description}
\item{programa}{programa}
\item{question_split}{question_split}
}
}
\usage{
data(data_sqi_mapper)
}
\description{
Quadro de referência para a recodificação dos nomes dos indicadores de supervisão e da chave de filtragem
}
\keyword{datasets}
