% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_sqi.R
\name{process_sqi_batch}
\alias{process_sqi_batch}
\title{Processar em lote o ficheiro de supervisão CSV}
\usage{
process_sqi_batch(path, var_mapping = data_sqi_mapper)
}
\arguments{
\item{path}{Caminho para a pasta de ficheiros onde o CSV esta guardado}

\item{var_mapping}{Tabela com mapeamento de variáveis de supervisão}
}
\value{
Quadro arrumado contendo dados de supervisão
}
\description{
Processar em lote o ficheiro de supervisão CSV
}
\examples{
 \dontrun{

 df <- process_sqi_batch("Data/")}
}
