% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_sqi.R
\name{process_sqi_file}
\alias{process_sqi_file}
\title{Arrumar ficheiros CSV das supervisões HIV extraídos do servidor ODK}
\usage{
process_sqi_file(path, safe = FALSE)
}
\arguments{
\item{path}{Caminho para a pasta de ficheiros onde o CSV esta guardado}

\item{safe}{Argumento opcional que, quando definido como \code{TRUE}, gera um relatório de erros detalhado}
}
\value{
Quadro arrumado contendo dados de supervisão
}
\description{
Arrumar ficheiros CSV das supervisões HIV extraídos do servidor ODK
}
\examples{
 \dontrun{

 df <- process_sqi_file("Data/")}
}
