% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector-package.R
\docType{package}
\name{inspector-package}
\alias{inspector}
\alias{inspector-package}
\title{inspector: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://usaid-mozambique.github.io/inspector/}
}

}
\author{
\strong{Maintainer}: Joe Lara \email{jlara@uw.edu}

}
\keyword{internal}
